// mssql/src/reactviews/pages/QueryResult/table/interfaces.ts
export interface FilterableColumn<T extends Slick.SlickData> extends Slick.Column<T> {
    filterable?: boolean;
    filterValues?: Array<string>;
    sorted?: SortProperties;
}